pyproject.toml
[tool.poetry]
name = "goals-service"
version = "0.1.0"
description = "Goals Service"
authors = ["Federico Pacheco <fpacheco@fi.uba.ar>", "Martina Panetta <mpanetta@fi.uba.ar>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.66.0" 
uvicorn = "^0.14.0"
pymongo = "^4.3.3"
flake8 = "^3.9.0"
black = "^22.3.0"
pytest = { version = "^6.2.5",  optional = true }
pytest-cov = { version = "^2.11.1", optional = true }
httpx = "^0.23.0"
requests = { version = "^2.28.2", optional = true }
pyjwt = "^2.7.0"
firebase-admin = "^6.1.0"
python-dateutil = "^2.8.2"

[tool.poetry.extras]
dev = ["pytest", "pytest-cov", "requests"]

[tool.black]
line-length = 88
skip-string-normalization = true